+++ src/forms/index.js
@@ -1,0 +1,4 @@
+const { form, Form } = require('./form');
+const { field, FieldDesriptor } = require('../services/fields');
+
+module.exports = { form, Form, field, FieldDesriptor };
